backend/requirements.txt
# Core
fastapi~=0.115.0
uvicorn[standard]~=0.34.0
python-multipart~=0.0.18
python-dotenv~=1.0.0

# AI
google-genai~=1.0.0

# Image Processing
Pillow~=11.0.0
piexif~=1.1.3
numpy~=1.26.0

# HTTP (for Photo Scout)
httpx~=0.27.0

# Retry logic
tenacity~=9.0.0

# Rate limiting
slowapi~=0.1.9
